// web/src/types/generated/types.gen.ts
export type PotentialStudentFilter = {
    /**
     * Limit
     *
     * Max objects
     */
    limit?: number;
    /**
     * Offset
     *
     * Offset
     */
    offset?: number;
    /**
     * Sort By
     *
     * Sort column
     */
    sort_by?: 'fullname' | 'created_date' | 'is_archived' | null;
    /**
     * Sort Desc
     *
     * Sort descending
     */
    sort_desc?: boolean;
    /**
     * Search Str
     *
     * Search by fullname and contacts
     */
    search_str?: string | null;
    /**
     * Is Archived
     *
     * Filter by archive
     */
    is_archived?: boolean | null;
    /**
     * Student Uid
     *
     * Filter by linked student
     */
    student_uid?: string | null;
    /**
     * Created By Uid
     *
     * Filter by creator
     */
    created_by_uid?: string | null;
    /**
     * Uid
     *
     * Filter by uid
     */
    uid?: string | null;
};

/**
 * student.PotentialStudentInput
 */
export type PotentialStudentInput = {
    /**
     * Fullname
     *
     * Full name
     */
    fullname: string;
    /**
     * Phone Number
     *
     * Phone number
     */
    phone_number?: string | null;
    /**
     * Telegram
     *
     * Telegram handle
     */
    telegram?: string | null;
    /**
     * Interests
     *
     * Student interests
     */
    interests?: string | null;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived?: boolean;
};

/**
 * student.PotentialStudentListedOutput
 */
export type PotentialStudentListedOutput = {
    /**
     * Items
     *
     * List of potential students
     */
    items: Array<PotentialStudentOutput>;
    /**
     * Count
     *
     * Total count
     */
    count: number;
};

/**
 * student.PotentialStudentOutput
 */
export type PotentialStudentOutput = {
    /**
     * Uid
     *
     * UID
     */
    uid: string;
    /**
     * Fullname
     *
     * Full name
     */
    fullname: string;
    /**
     * Phone Number
     *
     * Phone number
     */
    phone_number?: string | null;
    /**
     * Telegram
     *
     * Telegram handle
     */
    telegram?: string | null;
    /**
     * Interests
     *
     * Student interests
     */
    interests?: string | null;
    /**
     * Student Uid
     *
     * Linked student UID
     */
    student_uid?: string | null;
    /**
     * Created By Uid
     *
     * Creator user UID
     */
    created_by_uid: string;
    /**
     * Created Date
     *
     * Created date
     */
    created_date: string;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived: boolean;
};